requirements.txt
django>=5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
django-filter>=23.5
drf-spectacular>=0.27.0

celery>=5.3.0
redis>=5.0.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.0

psycopg2-binary>=2.9.9
dj-database-url>=2.1.0

python-decouple>=3.8
cryptography>=41.0.0
pydantic>=2.5.0

openai>=1.3.0
anthropic>=0.7.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-anthropic>=0.1.0
tiktoken>=0.5.0
django-redis>=5.4.0

markdown>=3.5.0
beautifulsoup4>=4.12.0
requests>=2.31.0

pytest>=7.4.0
pytest-django>=4.7.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
flake8>=6.1.0
black>=23.12.0
